Fix CI: eslint autofocus, Plotly sanitize export, and edge JSON API test.
Remove Auth autofocus eslint error, move sanitizePlotlyFigure out of the host module, harden bar-chart nulls and rule-tuning retries, and point the edge dashboard JSON API integration test at local /api/health instead of the public internet.
